+++ packages/common/component/BindI18n.vue
@@ -54,7 +54,7 @@
 <script>
 import { reactive, ref, watchEffect } from 'vue'
 import { useLayout, useTranslate } from '@opentiny/tiny-engine-controller'
-import { PROP_DATA_TYPE } from '@opentiny/tiny-engine-controller/utils'
+import { PROP_DATA_TYPE } from '@opentiny/tiny-engine-controller/js/constants'
 import { utils } from '@opentiny/tiny-engine-utils'
 import { Select, Option, Button, Input } from '@opentiny/vue'
 import { iconPlus } from '@opentiny/vue-icon'
