+++ packages/plugins/datasource/src/DataSourceRemotePanel.vue
@@ -48,8 +48,10 @@ import DataSrouceRemoteDataResult, { getResponseData } from './DataSourceRemoteD
 import { open as openRemoteMapping } from './DataSourceRemoteMapping.vue'
 import { useDataSource, useNotify } from '@opentiny/tiny-engine-controller'
 import { isEmptyObject } from '@opentiny/vue-renderless/common/type'
-import { reactiveObj2String as obj2String, string2Obj } from '@opentiny/tiny-engine-controller/utils'
+import { utils } from '@opentiny/tiny-engine-utils'
 import { getRequest } from './js/datasource'
+
+const { reactiveObj2String: obj2String, string2Obj } = utils
 
 export const isOpen = ref(false)
