+++ packages/settings/styles/src/components/layout/FlexBox.vue
@@ -23,11 +23,13 @@
 <script>
 import { ref, watchEffect } from 'vue'
 import { MetaRadio } from '@opentiny/tiny-engine-common'
-import { hyphenate } from '@opentiny/tiny-engine-controller/utils'
+import { utils } from '@opentiny/tiny-engine-utils'
 import ModalMask, { useModal } from '../inputs/ModalMask.vue'
 import ResetButton from '../inputs/ResetButton.vue'
 import { useProperties } from '../../js/useStyle'
 import { FLEX_PROPERTY } from '../../js/styleProperty'
+
+const { hyphenate } = utils
 
 export default {
   components: {
